HDFS-17435. Fix TestRouterRpc#testClearStaleNamespacesInRouterStateIdContext() failed
